refactor(cli): split frontend arguments parser into submodules
